requirements.txt
# Core Dependencies
adbutils==2.10.1
Pillow==10.0.1
python-dotenv==1.0.0
jupytext==1.15.1

# Distributed Setup Dependencies
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
requests==2.31.0

# Enhanced AI Integration
openai==1.3.7
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3

# Phi Ground Integration
transformers>=4.36.0
#torch>=2.1.0
#torchvision>=0.16.0
accelerate>=0.24.0
sentencepiece>=0.1.99
protobuf>=4.24.4
safetensors>=0.4.0
huggingface-hub>=0.19.0
#flash-attn==2.8.1

# Configuration Management
pydantic==2.4.2
pydantic-settings==2.0.3

# Advanced Logging
loguru==0.7.2

# Computer Vision & Image Processing
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-image==0.21.0

# OMniParser Integration (Enhanced UI Detection)
requests>=2.31.0
Pillow>=10.0.1

# Resource Monitoring & Performance
psutil==5.9.5
memory-profiler==0.60.0

# Async Support & Concurrency
asyncio-throttle==1.0.2
aiofiles==23.2.1

# Data Structures & Caching
redis==5.0.1
diskcache==5.6.3

# Testing & Quality Assurance
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Documentation & Utilities
rich==13.6.0
typer==0.9.0

# Version Control & Git Integration
GitPython==3.1.37

# Optional ML/RL Dependencies (for future phases)
gymnasium==0.29.1